LPS-194841 Inline class AssetRendererSearchResultInterpreter in class SearchResultInterpreterProvider
